requirements.txt
anyio==3.6.1
bcrypt==3.2.2
cffi==1.15.1
click==8.1.3
dnspython==2.2.1
fastapi==0.79.1
fastapi-csrf-protect==0.2.2
gunicorn==20.1.0
h11==0.13.0
idna==3.3
itsdangerous==2.1.2
motor==3.0.0
passlib==1.7.4
pycparser==2.21
pydantic==1.9.2
PyJWT==2.4.0
pymongo==4.2.0
python-decouple==3.6
sniffio==1.2.0
starlette==0.19.1
typing_extensions==4.3.0
uvicorn==0.18.2
